[Image]Change the return value of `pixel_size` to `u32`, since it is not necessary to use 64-bit interger to store the pixel size value.

+++ Modules/Luna/Image/Image.hpp
@@ -42,7 +42,7 @@ namespace Luna
 			rgba32_float,
 		};
 
-		inline constexpr usize pixel_size(ImagePixelFormat format)
+		inline constexpr u32 pixel_size(ImagePixelFormat format)
 		{
 			switch (format)
 			{
